Move VersionControl to it's own file
